Moved Chassis commands to chassis package instead of chassis.commands package and made all package protected.  Added factory in Chassis to make bugfixing of null pointers easier.
