+++ src/main/java/org/tahomarobotics/robot/OI.java
@@ -5,16 +5,10 @@
 
 package org.tahomarobotics.robot;
 
-import com.pathplanner.lib.auto.AutoBuilder;
-import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import org.tahomarobotics.robot.chassis.Chassis;
-import org.tahomarobotics.robot.chassis.commands.KnownMovementCommand;
-import org.tahomarobotics.robot.chassis.commands.TeleopDriveCommand;
 import org.tahomarobotics.robot.util.SubsystemIF;
 
 public class OI extends SubsystemIF {
@@ -53,16 +47,16 @@ public class OI extends SubsystemIF {
 //        Code for testing odometry
 //        driveController.x().onTrue(Commands.runOnce(chassis::zeroPose));
 //
-//        driveController.povLeft().onTrue(new KnownMovementCommand(0.5, 0.0, 0.0, p -> p.getTranslation().getX() < 2.0)
-//                .andThen(new KnownMovementCommand(0.0, 0.5, 0.0, p -> p.getTranslation().getY() < 2.0)
-//                .andThen(new KnownMovementCommand(-0.5, 0.0, 0.0, p -> p.getTranslation().getX() > 0.0)
-//                .andThen(new KnownMovementCommand(0.0, -0.5, 0.0, p -> p.getTranslation().getY() > 0.0)))));
+//        driveController.povLeft().onTrue(chassis.getKnownMovementCommand(0.5, 0.0, 0.0, p -> p.getTranslation().getX() < 2.0)
+//                .andThen(chassis.getKnownMovementCommand(0.0, 0.5, 0.0, p -> p.getTranslation().getY() < 2.0)
+//                .andThen(chassis.getKnownMovementCommand(-0.5, 0.0, 0.0, p -> p.getTranslation().getX() > 0.0)
+//                .andThen(chassis.getKnownMovementCommand(0.0, -0.5, 0.0, p -> p.getTranslation().getY() > 0.0)))));
     }
 
     private void setDefaultCommands() {
         Chassis chassis = Chassis.getInstance();
 
-        chassis.setDefaultCommand(new TeleopDriveCommand(
+        chassis.setDefaultCommand(chassis.getTeleopDriveCommand(
                 inputs -> {
                     inputs.x = -desensitizePowerBased(driveController.getLeftY(), FORWARD_SENSITIVITY);
                     inputs.y = -desensitizePowerBased(driveController.getLeftX(), FORWARD_SENSITIVITY);
